# backend/openapi/nomad.yaml
openapi: 3.0.3
info:
  title: NomadRide API
  description: API for the NomadRide intelligent motorcycle assistant.
  version: 1.0.0
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/fuel/cheapest:
    get:
      summary: Get cheapest and best value fuel stations
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [sp98, sp95, e10]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: lat
          in: query
          required: false
          schema:
            type: number
        - name: lon
          in: query
          required: false
          schema:
            type: number
        - name: radius
          in: query
          required: false
          schema:
            type: number
            default: 15
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: fillSize
          in: query
          required: false
          schema:
            type: number
            default: 9
        - name: consumption
          in: query
          required: false
          schema:
            type: number
            default: 4
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /api/fuel/autocomplete:
    get:
      summary: Auto-complete location queries using Nominatim
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    lat:
                      type: number
                    lon:
                      type: number